Create util.js a js file with functions that are common to all other js files
