+++ portfolio/src/main/webapp/index.js
@@ -12,15 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {removeClassName, displayDefaultSection, displaySection, setToNone} from "./tab.js";
-import {showNextSlide, showSlide, showCurrentSlide, slideIndex} from "./slideshow.js";
+import {setToNone} from "./utils.js";
+import {displayDefaultSection, displaySection} from "./tab.js";
+import {showCurrentSlide, showNextSlide, showSlide, slideIndex} from "./slideshow.js";
 
 /** Allow all imported functions to be in the scope of "window". */
 window.onload = () => {
   window.displaySection = displaySection;
-  window.setToNone = setToNone;
   window.displayDefaultSection = displayDefaultSection;
-  window.removeClassName = removeClassName;
   window.showNextSlide = showNextSlide;
   window.showSlide = showSlide;
   window.showCurrentSlide = showCurrentSlide;
