+++ portfolio/src/main/webapp/tab.js
@@ -11,6 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+import {removeClassName, setToNone} from "./utils.js";
 
 /**  
  * Sets the default-tab by "clicking" on the "Background" tab, 
@@ -44,30 +45,4 @@ const displaySection = (event, sectionName) => {
   event.currentTarget.classList.add("active");
 }
 
-/** 
- * Set all elements of given className name to style.display = none.
- * @param {string} name - The name of the class that I need to modify.
- */
-const setToNone = (name) => {
-  const content = document.getElementsByClassName(name);
-  for (let i = 0; i < content.length; i++) {
-    content[i].classList.remove("default-block");
-    content[i].classList.add("default-none");
-  }
-}
-
-/**
- * Sets className of all elements of a given 
- * className name from "active" to default "".
- * @param {string} name - The name of the class I need to modify.
- * @param {string} toRemove - The string that I need to remove 
- *  from the class name I'm modifying.
- */
-const removeClassName = (name, toRemove) => {
-  const elements = document.getElementsByClassName(name);
-  for (let i = 0; i < elements.length; i++) {
-    elements[i].classList.remove(toRemove);
-  }
-}
-
-export {displayDefaultSection, displaySection, setToNone, removeClassName};
+export {displayDefaultSection, displaySection};
